spring configuration changes to make the rest api changes and ui changes backwards compatable.
moving the seedController and seedFormController into one controller and handling their logic in two respective services. NOTE: formService is posting empty data right now, don't understand why.

fixed empty post from ui.
