+++ geowebcache/rest/src/main/java/org/geowebcache/rest/controller/SeedController.java
@@ -23,27 +23,12 @@
 
 package org.geowebcache.rest.controller;
 
-import com.google.common.base.Splitter;
-import com.thoughtworks.xstream.XStream;
-import com.thoughtworks.xstream.io.HierarchicalStreamDriver;
-import com.thoughtworks.xstream.io.HierarchicalStreamReader;
-import com.thoughtworks.xstream.io.copy.HierarchicalStreamCopier;
-import com.thoughtworks.xstream.io.json.JettisonMappedXmlDriver;
-import com.thoughtworks.xstream.io.json.JsonHierarchicalStreamDriver;
-import com.thoughtworks.xstream.io.xml.DomDriver;
-import com.thoughtworks.xstream.io.xml.PrettyPrintWriter;
-import org.apache.commons.io.IOUtils;
 import org.geowebcache.GeoWebCacheException;
-import org.geowebcache.config.ContextualConfigurationProvider;
 import org.geowebcache.config.XMLConfiguration;
-import org.geowebcache.io.GeoWebCacheXStream;
-import org.geowebcache.layer.TileLayer;
 import org.geowebcache.rest.exception.RestException;
-import org.geowebcache.seed.GWCTask;
-import org.geowebcache.seed.SeedRequest;
+import org.geowebcache.rest.service.FormService;
+import org.geowebcache.rest.service.SeedService;
 import org.geowebcache.seed.TileBreeder;
-import org.json.JSONException;
-import org.json.JSONObject;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -51,13 +36,6 @@ import org.springframework.stereotype.Component;
 import org.springframework.web.bind.annotation.*;
 
 import javax.servlet.http.HttpServletRequest;
-import java.io.IOException;
-import java.io.StringReader;
-import java.io.StringWriter;
-import java.util.Iterator;
-import java.util.LinkedList;
-import java.util.List;
-import java.util.Map;
 
 @Component
 @RestController
@@ -66,6 +44,12 @@ public class SeedController {
 
     @Autowired
     TileBreeder seeder;
+
+    @Autowired
+    SeedService seedService;
+
+    @Autowired
+    FormService formService;
 
     @Autowired
     protected XMLConfiguration xmlConfig;
@@ -82,16 +66,7 @@ public class SeedController {
      */
     @RequestMapping(value = "/seed.json", method = RequestMethod.GET)
     public ResponseEntity<?> doGet(HttpServletRequest req) {
-        try {
-            XStream xs = new GeoWebCacheXStream(new JsonHierarchicalStreamDriver());
-            JSONObject obj = null;
-            long[][] list;
-            list = seeder.getStatusList();
-            obj = new JSONObject(xs.toXML(list));
-            return new ResponseEntity<>(obj.toString(), HttpStatus.OK);
-        } catch (JSONException jse) {
-            return new ResponseEntity<Object>("error", HttpStatus.INTERNAL_SERVER_ERROR);
-        }
+        return seedService.getRunningTasks(req);
     }
 
     /**
@@ -100,35 +75,20 @@ public class SeedController {
      * @param layer
      * @return
      */
-    @RequestMapping(value = "/seed/{layer:.+}", method = RequestMethod.GET)
+    @RequestMapping(value = "/seed/{layer}.json", method = RequestMethod.GET)
     public ResponseEntity<?> doGet(HttpServletRequest req, @PathVariable String layer) {
-        final String layerName;
-        if (layer.indexOf(".") != -1) {
-            layerName = layer.substring(0, layer.indexOf("."));
-        } else {
-            layerName = layer;
-        }
+        return seedService.getRunningLayerTasks(req, layer);
+    }
 
-        try {
-            XStream xs = new GeoWebCacheXStream(new JsonHierarchicalStreamDriver());
-            JSONObject obj = null;
-            long[][] list;
-            if (null == layerName) {
-                list = seeder.getStatusList();
-            } else {
-                try {
-                    seeder.findTileLayer(layerName);
-                } catch (GeoWebCacheException e) {
-                    return new ResponseEntity<String>(e.getMessage(), HttpStatus.BAD_REQUEST);
-                }
-                list = seeder.getStatusList(layerName);
-            }
-            obj = new JSONObject(xs.toXML(list));
-            return new ResponseEntity<>(obj.toString(), HttpStatus.OK);
-        } catch (JSONException jse) {
-            jse.printStackTrace();
-            return new ResponseEntity<Object>("error", HttpStatus.INTERNAL_SERVER_ERROR);
-        }
+    /**
+     * GET method for displaying the GeoWebCache UI form.
+     * @param request
+     * @param layer
+     * @return
+     */
+    @RequestMapping(value = "/seed/{layer}", method = RequestMethod.GET)
+    public ResponseEntity<?> doFormGet(HttpServletRequest request, @PathVariable String layer) {
+        return formService.handleGet(request, layer);
     }
 
     /**
@@ -138,7 +98,7 @@ public class SeedController {
      */
     @RequestMapping(value = "/seed", method = RequestMethod.POST)
     public ResponseEntity doPost(HttpServletRequest request) {
-        String response = handleKillAllThreads(request, null);
+        String response = seedService.handleKillAllThreads(request, null);
         if (response.equalsIgnoreCase("error")) {
             return new ResponseEntity(HttpStatus.INTERNAL_SERVER_ERROR);
         } else {
@@ -147,230 +107,26 @@ public class SeedController {
     }
 
     /**
-     * POST method to kill [all, running, pending] tasks for a layer
-     *
-     * This request mapping also catches the POST for seeding and truncating
-     * due to the way Spring handles stripping out extensions when using <mvc:annotation-driven />
-     * this feature can be overwritten but for now will check the extension of the request
-     * and if it is xml or json then we will call the seeding method otherwise we will kill tasks.
-     *
-     * We are using a regex to force spring to keep any extensions
-     * @param request
-     * @return
-     */
-    @RequestMapping(value = "/seed/{layer:.+}", method = RequestMethod.POST)
-    public ResponseEntity doPost(HttpServletRequest request, @PathVariable String layer) {
-        if (layer.indexOf(".") != -1) {
-            String layerName = layer.substring(0, layer.indexOf("."));
-            String extension = layer.substring(layer.indexOf(".") + 1);
-            return doSeeding(request, extension, layerName);
-        } else {
-            String response = handleKillAllThreads(request, layer);
-            if (response.equalsIgnoreCase("error")) {
-                return new ResponseEntity(HttpStatus.INTERNAL_SERVER_ERROR);
-            } else {
-                return new ResponseEntity<String>(response, HttpStatus.OK);
-            }
-        }
-    }
-
-    /**
-     * Internal method to do the seeding and truncating until the issue with spring
-     * stripping extensions when mapping Requests can be resolved.
-     * @param request
-     * @param extension
-     * @param layer
-     * @return
-     */
-    private ResponseEntity<?> doSeeding(HttpServletRequest request, String extension, String layer) {
-        XStream xs = configXStream(new GeoWebCacheXStream(new DomDriver()));
-
-        Object obj = null;
-
-        try {
-            StringWriter writer = new StringWriter();
-            IOUtils.copy(request.getInputStream(), writer, "UTF-8");
-            String requestData = writer.toString();
-
-            if (extension==null || extension.equalsIgnoreCase("xml")) {
-                obj = xs.fromXML(requestData);
-            } else if (extension.equalsIgnoreCase("json")) {
-                obj = xs.fromXML(convertJson(requestData));
-            } else {
-                throw new RestException("Format extension unknown or not specified: "
-                        + extension, HttpStatus.BAD_REQUEST);
-            }
-            handleRequest(layer, obj);
-            return new ResponseEntity<Object>(HttpStatus.OK);
-        } catch (IOException e) {
-            return new ResponseEntity<Object>(HttpStatus.INTERNAL_SERVER_ERROR);
-        }
-    }
-
-    /**
      * POST method for Seeding and Truncating
      * @param request
      * @param layer
-     * @param extension
      * @return
      */
-    @RequestMapping(value = "/seed/{layer}.{extension}", method = RequestMethod.POST)
+    @RequestMapping(value = "/seed/{layer:.+}", method = RequestMethod.POST)
     public ResponseEntity<?> doPost(HttpServletRequest request,
-                                    @PathVariable String layer,
-                                    @PathVariable String extension) {
-        XStream xs = configXStream(new GeoWebCacheXStream(new DomDriver()));
-
-        Object obj = null;
-
-        try {
-            StringWriter writer = new StringWriter();
-            IOUtils.copy(request.getInputStream(), writer, "UTF-8");
-            String requestData = writer.toString();
-
-            if (extension==null || extension.equalsIgnoreCase("xml")) {
-                obj = xs.fromXML(requestData);
-            } else if (extension.equalsIgnoreCase("json")) {
-                obj = xs.fromXML(convertJson(requestData));
-            } else {
-                throw new RestException("Format extension unknown or not specified: "
-                        + extension, HttpStatus.BAD_REQUEST);
-            }
-            handleRequest(layer, obj);
-            return new ResponseEntity<Object>(HttpStatus.OK);
-        } catch (IOException e) {
-            return new ResponseEntity<Object>(HttpStatus.INTERNAL_SERVER_ERROR);
-        }
-    }
-
-    /**
-     * METHOD that handles the seeding/truncating task from the POST method.
-     * @param layerName
-     * @param obj
-     */
-    protected void handleRequest(String layerName, Object obj) {
-        final SeedRequest sr = (SeedRequest) obj;
-        try {
-            seeder.seed(layerName, sr);
-        } catch (IllegalArgumentException e) {
-            throw new RestException(e.getMessage(), HttpStatus.BAD_REQUEST);
-        } catch (GeoWebCacheException e) {
-            throw new RestException(e.getMessage(), HttpStatus.INTERNAL_SERVER_ERROR);
-        }
-    }
-
-    /**
-     * Deserializing a json string is more complicated.
-     *
-     * XStream does not natively support it. Rather, it uses a JettisonMappedXmlDriver to convert to
-     * intermediate xml and then deserializes that into the desired object. At this time, there is a
-     * known issue with the Jettison driver involving elements that come after an array in the json
-     * string.
-     *
-     * http://jira.codehaus.org/browse/JETTISON-48
-     *
-     * The code below is a hack: it treats the json string as text, then converts it to the
-     * intermediate xml and then deserializes that into the SeedRequest object.
-     */
-    protected String convertJson(String entityText) throws IOException {
-        HierarchicalStreamDriver driver = new JettisonMappedXmlDriver();
-        StringReader reader = new StringReader(entityText);
-        HierarchicalStreamReader hsr = driver.createReader(reader);
-        StringWriter writer = new StringWriter();
-        new HierarchicalStreamCopier().copy(hsr, new PrettyPrintWriter(writer));
-        writer.close();
-        return writer.toString();
-    }
-
-    /**
-     * Splits a string into a K,V Map
-     * @param in
-     * @return
-     */
-    private Map<String, String> splitToMap(String in) {
-        return Splitter.on(" ").withKeyValueSeparator("=").split(in);
-    }
-
-    protected XStream configXStream(XStream xs) {
-        return xmlConfig.getConfiguredXStreamWithContext(xs, ContextualConfigurationProvider.Context.REST);
-    }
-
-    public void setTileBreeder(TileBreeder seeder) {
-        this.seeder = seeder;
-    }
-
-    /**
-     * Method to kill running tasks for all of GWC or just the provided layer.
-     * @param request
-     * @param layer
-     * @return
-     */
-    private String handleKillAllThreads(HttpServletRequest request, String layer) {
-        final TileLayer tl;
-        if (layer != null) {
+                                    @PathVariable String layer) {
+        if (layer.indexOf(".") == -1) {
             try {
-                tl = seeder.findTileLayer(layer);
+                return formService.handleFormPost(request, layer);
             } catch (GeoWebCacheException e) {
-                throw new RestException(e.getMessage(), HttpStatus.BAD_REQUEST);
+                return new ResponseEntity<Object>("error", HttpStatus.INTERNAL_SERVER_ERROR);
             }
+
         } else {
-            tl = null;
-        }
-        try {
-            StringBuilder doc = new StringBuilder();
-            StringWriter writer = new StringWriter();
-            IOUtils.copy(request.getInputStream(), writer, "UTF-8");
-            String body = writer.toString();
-            Map<String, String> commandMap = splitToMap(body);
-
-            final boolean allLayers = tl == null;
-
-            if (commandMap.containsKey("kill_all")) {
-                String killCode = commandMap.get("kill_all");
-                final Iterator<GWCTask> tasks;
-                if ("1".equals(killCode) || "running".equalsIgnoreCase(killCode)) {
-                    killCode = "running";
-                    tasks = seeder.getRunningTasks();
-                } else if ("pending".equalsIgnoreCase(killCode)) {
-                    tasks = seeder.getPendingTasks();
-                } else if ("all".equalsIgnoreCase(killCode)) {
-                    tasks = seeder.getRunningAndPendingTasks();
-                } else {
-                    throw new RestException("Unknown kill_all code: '" + killCode
-                            + "'. One of all|running|pending is expected.", HttpStatus.BAD_REQUEST);
-                }
-                List<GWCTask> terminatedTasks = new LinkedList<GWCTask>();
-                List<GWCTask> nonTerminatedTasks = new LinkedList<GWCTask>();
-                while (tasks.hasNext()) {
-                    GWCTask task = tasks.next();
-                    String layerName = task.getLayerName();
-                    if (!allLayers && !tl.getName().equals(layerName)) {
-                        continue;
-                    }
-                    long taskId = task.getTaskId();
-                    boolean terminated = seeder.terminateGWCTask(taskId);
-                    if (terminated) {
-                        terminatedTasks.add(task);
-                    } else {
-                        nonTerminatedTasks.add(task);
-                    }
-                }
-
-                doc.append("<p>Requested to terminate ").append(killCode).append(" tasks.");
-                doc.append("Terminated tasks: <ul>");
-                for (GWCTask t : terminatedTasks) {
-                    doc.append("<li>").append(t).append("</li>");
-                }
-                doc.append("</ul>Tasks already finished: <ul>");
-                for (GWCTask t : nonTerminatedTasks) {
-                    doc.append("<li>").append(t).append("</li>");
-                }
-                if (tl != null) {
-                    doc.append("</ul><p><a href=\"./" + tl.getName() + "\">Go back</a></p>\n");
-                }
-            }
-            return doc.toString();
-        } catch (IOException e) {
-            return "error";
+            String extension = layer.substring(layer.indexOf(".") +1);
+            String layerName = layer.substring(0, layer.indexOf("."));
+            return seedService.doSeeding(request, layerName, extension);
         }
     }
+
 }
