+++ geowebcache/web/src/main/webapp/WEB-INF/geowebcache-rest-context.xml
@@ -7,7 +7,6 @@
     http://www.springframework.org/schema/context http://www.springframework.org/schema/context/spring-context-4.0.xsd
     http://www.springframework.org/schema/mvc http://www.springframework.org/schema/mvc/spring-mvc-4.0.xsd">
 
-  <mvc:annotation-driven />
   <context:component-scan base-package="org.geowebcache.rest, org.geowebcache.diskquota.rest.controller" />
 
 </beans>
